Finish removing text experiment global enable
