+++ Source/Base/ASAvailability.h
@@ -57,6 +57,10 @@
   #define YOGA __has_include(YOGA_HEADER_PATH)
 #endif
 
+#ifdef ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE
+  #error "ASTEXTNODE_EXPERIMENT_GLOBAL_ENABLE is unavailable. See ASConfiguration.h."
+#endif
+
 #define AS_PIN_REMOTE_IMAGE __has_include(<PINRemoteImage/PINRemoteImage.h>)
 #define AS_IG_LIST_KIT __has_include(<IGListKit/IGListKit.h>)
